THE CHAMELEON AGENT/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM Providers
google-generativeai==0.3.2
groq==0.4.1

# NLP & Entity Extraction
spacy==3.7.2
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
locust==2.20.0

# Utilities
python-multipart==0.0.6
